requirements.txt
accelerate==0.27.2
addict==2.4.0
asttokens==2.4.1
backcall==0.2.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
diffusers==0.26.3
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.1.1
gpustat==1.1.1
# Editable install with no version control (groundingdino==0.1.0)
-e ./models/GroundingDINO
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.3
loguru==0.7.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.535.133
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prefetch_generator==1.0.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
supervision==0.3.2
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yapf==0.40.2
zipp==3.17.0
